requirements.txt
pyTelegramBotAPI
flask
requests
beautifulsoup4
gunicorn
psycopg2-binary